defined variables not defined before
matkeep was just the mat12 variable. "pheno" and "p" were not defined assuming one phenotype was going to be run at a time. I've added the vector of phenotypes in "pheno" and loop through them using the variable "p" just as I had run them. But "pheno" can be defined as a string with one phenotype and the variable "p" can be removed from the association being run in (now) line 78.